Validation for user to do actions

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import { NextResponse } from 'next/server';
 import * as yup from 'yup';
 
@@ -37,11 +38,16 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+    const user = await getUserSessionServer();
+    if ( !user ) {
+        return NextResponse.json('No autorizado', { status: 401 });
+    }
+
     try {
             
         const { complete, description } = await postSchema.validate(await request.json());
         
-        const todo = await prisma?.todo.create({ data: { complete, description } });
+        const todo = await prisma?.todo.create({ data: { complete, description, userID: user.id } });
         
         return NextResponse.json(todo);
         
@@ -53,9 +59,14 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) {
 
+    const user = await getUserSessionServer();
+    if ( !user ) {
+        return NextResponse.json('No autorizado', { status: 401 });
+    }
+
     try {
             
-        await prisma?.todo.deleteMany({ where:{ complete: false }});
+        await prisma?.todo.deleteMany({ where:{ complete: false, userID: user.id }});
         
         return NextResponse.json({
             message: 'Todos borrados'
